Implement createApiKey (#2)
createApiKey generates a public/private key pair, a uuid key id.
It modifies the issuerUrl to have the kid as the suffix. Then, it takes all the claims and signs the JWT with the given claims. Finally, it converts the public key to JWK format and returns the JWT and the JWKS

This commit also has some misc improvements:
* Implement coverage
* Restrict to node 19+ to make sure crypto is defined