Simplify tsconfig to extend from a root config (#13)
Each sub-package needs to re-specify baseUrl to reset the paths
This additionally fixes it so type changes are reflected
immediately in development, without requiring npm run build
The downside is that in order for this to work, the module type
has to be ESNext, not NodeNext. For libraries that target nodejs
technically this could emit erroneous outputs, or miss type checking
issues.

However, for the cases we're interested in, especially since we're
using other flags such as rewritableRelativeImportExtensions
it should be okay